chat add more fet

+++ lib/features/messages/data/data_sources_impl/remote/delete_message_remote_data_source_impl/delete_message_remote_data_source_impl.dart
@@ -4,6 +4,8 @@ import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:taskly/core/errors/failures.dart';
 import 'package:taskly/core/services/supabase_service.dart';
 import 'package:taskly/features/messages/data/data_sources/remote/delete_message_remote_data_source/delete_message_remote_data_source.dart';
+
+import '../../../../../../core/utils/network_utils.dart';
 @Injectable(as: DeleteMessageRemoteDataSource)
 class DeleteMessageRemoteDataSourceImpl extends  DeleteMessageRemoteDataSource{
   final SupabaseService supabaseService;
@@ -11,6 +13,9 @@ class DeleteMessageRemoteDataSourceImpl extends  DeleteMessageRemoteDataSource{
    @override
   Future<Either<Failures, void>> deleteMessage(String messageId) async {
     try {
+      if (!await NetworkUtils.hasInternet()) {
+        return const Left(NetworkFailure('تحقق من اتصالك بالانترنت'));
+      }
       await supabase
           .from('messages')
           .delete()
